test(unit): converted the tests for the documentation scaffolder to vitest
for #452
